+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Global. Hover effect on Android devices (expected: no-hover on touch-screens)
 - Global. Focus frame isn't rouned on Safari
 - Combobox based controls (Select, Date, Time). Now popup isn't open if user clears control and it gets focus at the same time
+- [TextareaControl](src/controls/textarea.ts). `Ctrl + B` makes text bold but it's unexpected for plain textarea
 
 - **New**: